Removed comments,spaces. Separate footer-mortgage
