+++ web/rainmaker/dev-packages/egov-rented-properties-dev/src/ui-config/screens/specs/rented-properties-citizen/footer-mortgage.js
@@ -1,0 +1,148 @@
+import { getCommonApplyFooter, validateFields } from "../utils";
+import { toggleSnackbar } from "egov-ui-framework/ui-redux/screen-configuration/actions";
+import get from "lodash/get";
+import { applyOwnershipTransfer, getDetailsFromProperty } from "../../../../ui-utils/apply";
+import { previousButton, submitButton, nextButton, changeStep, moveToSuccess, DETAILS_STEP, DOCUMENT_UPLOAD_STEP, SUMMARY_STEP } from "../rented-properties/applyResource/footer";
+import { some } from "lodash";
+import { prepareFinalObject } from "egov-ui-framework/ui-redux/screen-configuration/actions";
+
+
+
+const callBackForNextMortgage = async(state, dispatch) => {
+  let activeStep = get(
+      state.screenConfiguration.screenConfig["mortage-apply"],
+      "components.div.children.stepper.props.activeStep",
+      0
+  );
+  let isFormValid = true;
+  let hasFieldToaster = true;
+  if(activeStep === DETAILS_STEP) {
+    const isOwnerDetailsValid = validateFields(
+      "components.div.children.formwizardFirstStep.children.applicantDetails.children.cardContent.children.detailsContainer",            
+      state,
+      dispatch,
+      "mortage-apply"
+    )
+    const isAddressDetailsValid = validateFields(
+      "components.div.children.formwizardFirstStep.children.ownershipAddressDetails.children.cardContent.children.detailsContainer",            
+      state,
+      dispatch,
+      "mortage-apply"
+    )
+    if(!!isOwnerDetailsValid && !!isAddressDetailsValid) {
+            isFormValid = true;
+        } 
+    else {
+            isFormValid = false;
+        }
+}
+  if(activeStep === DOCUMENT_UPLOAD_STEP) {
+  //   const uploadedDocData = get(
+  //     state.screenConfiguration.preparedFinalObject,
+  //     "Properties[0].propertyDetails.applicationDocuments",
+  //     []
+  // );
+
+  // const uploadedTempDocData = get(
+  //     state.screenConfiguration.preparedFinalObject,
+  //     "PropertiesTemp[0].applicationDocuments",
+  //     []
+  // );
+
+  // for (var y = 0; y < uploadedTempDocData.length; y++) {
+  //   if (
+  //       uploadedTempDocData[y].required &&
+  //       !some(uploadedDocData, { documentType: uploadedTempDocData[y].name })
+  //   ) {
+  //       isFormValid = false;
+  //   }
+  // }
+  // if(isFormValid) {
+  //   const reviewDocData =
+  //           uploadedDocData &&
+  //           uploadedDocData.map(item => {
+  //               return {
+  //                   title: `RP_${item.documentType}`,
+  //                   link: item.fileUrl && item.fileUrl.split(",")[0],
+  //                   linkText: "View",
+  //                   name: item.fileName
+  //               };
+  //           });
+  //           dispatch(
+  //             prepareFinalObject("PropertiesTemp[0].reviewDocData", reviewDocData)
+  //         );
+  // }
+      isFormValid = true;
+  }
+
+  if(activeStep === SUMMARY_STEP) {
+  isFormValid = true;
+    if (isFormValid) {
+      const rentedData = get(
+        state.screenConfiguration.preparedFinalObject,
+        "Properties[0]"
+    );
+        moveToSuccess(rentedData, dispatch);
+    }
+    
+  }
+
+  if(activeStep !== SUMMARY_STEP) {
+      if (isFormValid) {
+        
+          changeStep(state, dispatch, "mortage-apply");
+      } else if (hasFieldToaster) {
+          let errorMessage = {
+              labelName:
+                  "Please fill all mandatory fields and upload the documents !",
+              labelKey: "ERR_FILL_MANDATORY_FIELDS_UPLOAD_DOCS"
+          };
+          switch (activeStep) {
+              case DETAILS_STEP:
+                  errorMessage = {
+                      labelName:
+                          "Please fill all mandatory fields, then do next !",
+                      labelKey: "ERR_FILL_RENTED_MANDATORY_FIELDS"
+                  };
+                  break;
+              case DOCUMENT_UPLOAD_STEP:
+                  errorMessage = {
+                      labelName: "Please upload all the required documents !",
+                      labelKey: "ERR_UPLOAD_REQUIRED_DOCUMENTS"
+                  };
+                  break;
+          }
+          dispatch(toggleSnackbar(true, errorMessage, "warning"));
+      }
+  }
+}
+
+
+export const callBackForPreviousMortgage = (state, dispatch) => {
+  changeStep(state, dispatch, "mortage-apply", "previous");
+};
+
+
+export const mortgagefooter = getCommonApplyFooter({
+  previousButton: {
+    ...previousButton,
+    onClickDefination: {
+      action: "condition",
+      callBack: callBackForPreviousMortgage
+    },
+  },
+  nextButton: {
+    ...nextButton,
+    onClickDefination: {
+      action: "condition",
+      callBack: callBackForNextMortgage
+    }
+  },
+  submitButton: {
+    ...submitButton,
+    onClickDefination: {
+      action: "condition",
+      callBack: callBackForNextMortgage
+    },
+  }
+});
