+++ web/rainmaker/dev-packages/egov-rented-properties-dev/src/ui-config/screens/specs/rented-properties-citizen/mortage-apply.js
@@ -3,7 +3,7 @@ import {
   } from "egov-ui-framework/ui-config/screens/specs/utils";
 
   import {stepper, formwizardMortgageFirstStep,formwizardMortgageSecondStep, formwizardMortgageThirdStep } from '../rented-properties/applyResource/applyConfig';
-  import {mortgagefooter} from './footer';
+  import {mortgagefooter} from './footer-mortgage';
 
 
 const header = getCommonHeader({
